% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_ipea.R
\encoding{UTF-8}
\name{codes_ipea}
\alias{codes_ipea}
\title{Get vailable series from IPEADATA}
\usage{
codes_ipea(language = c("en", "br"))
}
\arguments{
\item{language}{String specifying the selected language. Language options are English ("en", default) and Brazilian portuguese ("br").}
}
\value{
Tibble containing Ipeadata code, name, theme, source, frequency, last update and activity status of available series.
}
\description{
Returns a tibble with available series from IPEADATA'S API.
}
\examples{
all_codes <- codes_ipea()
}
