% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipea.R
\encoding{UTF-8}
\name{get_ipea}
\alias{get_ipea}
\title{Get IPEADATA series}
\usage{
get_ipea(code)
}
\arguments{
\item{code}{Character vector with series code.}
}
\value{
A tibble with data from IPEADATA
}
\description{
This function provides the extraction of data using IPADATA's API \url{http://www.ipeadata.gov.br/api/}.
}
\details{
See the \code{\link[=codes_ipea]{codes_ipea()}} to available series codes and other information.
}
\examples{
get_ipea(c("CAGED12_SALDON12", "CAGED12_DESLIGN12"))
}
